changed determineError() to discourage crash
Formerly patternError was set to null, and then immediately iterated upon which crashes 100% of the time. New version crashes less often, and sets patternError to the return value of calculatePatternError() which is probably the intended behavior. getErrorFunction() in SupervisedLearning returns an ErrorFunction. Use of ErrorFunction.reset() calculatePatternError() and getTotalError() may be advised in the future for the sake of consistency/code reuse.